Overlays with hue added to drawAllsky. Lines between tiles in overlays fixed.
Lines between overlays were caused by applying the alpha during drawing the tiles on bCtx and then again when drawing the bCtx canvas on the ctx canvas. The alpha value when drawing tiles is now set to 1.0. Therefore, the base layer is always fully opaque. If this is undesirable a workaround will need to be implemented to apply the alpha during drawOneTile2 but only for the base layer.