Images are now only sent to the DESonic app once panning and zooming has stopped.
This was implemented using a setTimeout to see if the zoom or coordinates has changed after 200ms (in updateImageAtIndex). This delay can be adjusted. Longer will mean that there is more delay between when the user stops zooming/panning and the image sending but shorter will mean that images will continue to send if the user is zooming slowly (zoom seems to be more quantised than panning which causes lots of continuous small changes in position).